% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb-package.R
\docType{package}
\name{tcdsb-package}
\alias{tcdsb}
\alias{tcdsb-package}
\title{tcdsb: TCDSB - Visual Identity}
\description{
This packae is to support visual identity for Toronto Catholic District School Board.
}
\author{
\strong{Maintainer}: Rousell Greg \email{grousell@gmail.com}

}
\keyword{internal}
