% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_colours_fonts.R
\name{tcdsb_colours_fonts}
\alias{tcdsb_colours_fonts}
\title{TCDSB Colors and Fonts}
\usage{
tcdsb_colours_fonts(...)
}
\arguments{
\item{...}{Not used. Included for possible future use.}
}
\value{
R objects with names and HEX colors for TCDSB visual identity
}
\description{
TCDSB Colors and Fonts
}
\examples{
# theme <- tcdsb_colours_fonts()
# theme$tcdsb_board_color
# "#951B1E"
}
