% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_diverge_plot_function.R
\name{tcdsb_diverge_plot_function}
\alias{tcdsb_diverge_plot_function}
\title{TCDSB Diverging Plots
A function that creates a diverging bar chart, using 2, 4 or 6 factor levels.}
\usage{
tcdsb_diverge_plot_function(
  df,
  group_col,
  response_col,
  fct_levels,
  percent_col = percent,
  label_width = 30,
  font = "Century Gothic",
  low_colour = "#8EB8C2",
  high_colour = "#d2a940",
  font_size = 1,
  left_label = "Left Label",
  right_label = "Right Label",
  bar_width = 0.75
)
}
\arguments{
\item{df}{Data frame in long format}

\item{group_col}{Column with groups}

\item{response_col}{Column with responses}

\item{fct_levels}{Factor levels}

\item{percent_col}{Column with percentages}

\item{label_width}{Width of labels}

\item{font}{Font}

\item{low_colour}{Low end colour}

\item{high_colour}{High end colour}

\item{font_size}{Font size scale - base is 1}

\item{left_label}{Left hand label}

\item{right_label}{Right hand label}

\item{bar_width}{Width of bars}
}
\value{
A ggplot object
}
\description{
TCDSB Diverging Plots
A function that creates a diverging bar chart, using 2, 4 or 6 factor levels.
}
\examples{
# Not run
# df |>
#   tcdsb_diverge_plot_function(
#     group_col = Question,
#     response_col = response,
#     percent_col = percent,
#     fct_levels = c("Strongly Disagree", "Disagree", "Agree", "Strongly Agree"))
}
