% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_diverge_plot_other_function.R
\name{tcdsb_diverge_plot_other_function}
\alias{tcdsb_diverge_plot_other_function}
\title{TCDSB Diverging Plots - Other
A function to create plots of other responses (i.e. Missing, Neutral) to pair
with TCDSB Diverging Plots}
\usage{
tcdsb_diverge_plot_other_function(
  df,
  group_col,
  response_col,
  percent_col,
  label,
  font = "Century Gothic",
  bar_color = "grey70",
  label_width = 30,
  legend_width = 10,
  x_max = 1,
  font_size = 1,
  bar_width = 0.75
)
}
\arguments{
\item{df}{Data frame in long format}

\item{group_col}{Column with groups}

\item{response_col}{Column with responses}

\item{percent_col}{Column with percentages}

\item{label}{Label for chart}

\item{font}{Font}

\item{bar_color}{Bar color (defaults to grey)}

\item{label_width}{Width of labels}

\item{legend_width}{Width of legend}

\item{x_max}{Max of x-axis. Can be adjusted for appearance}

\item{font_size}{Font size scale - base is 1}

\item{bar_width}{Width of bars}
}
\value{
A ggplot to be paired with tcdsb_diverge_plot
}
\description{
TCDSB Diverging Plots - Other
A function to create plots of other responses (i.e. Missing, Neutral) to pair
with TCDSB Diverging Plots
}
\examples{
# Not run
# df |>
#   dplyr::filter(response == "Neutral") |>
#   tcdsb_diverge_plot_other_function(
#     group_col = Question,
#     response_col = response,
#     percent_col = percent,
#     label = "Neutral"
#     )
}
