% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_fetch_alchemer_data.R
\name{tcdsb_fetch_alchemer_data}
\alias{tcdsb_fetch_alchemer_data}
\title{TCDSB Fetchs Alchemer Survey Data
A function that retrieves data from TCDSB's Alchemer account.
Requires user to have set up an API with Alchemer, including
a token and secret key}
\usage{
tcdsb_fetch_alchemer_data(survey_number)
}
\arguments{
\item{survey_number}{Alchemer survey number, retrieved from survey URL}
}
\value{
Cleaned data frame of survey data
}
\description{
TCDSB Fetchs Alchemer Survey Data
A function that retrieves data from TCDSB's Alchemer account.
Requires user to have set up an API with Alchemer, including
a token and secret key
}
\examples{
# token <- keyring::key_get("alchemer", "token")
# secret_key <- keyring::key_get("alchemer", "secret")
# get_alchemer_data("XXXXXXXX")
}
