% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_fetch_alchemer_data_dictionary.R
\name{tcdsb_fetch_alchemer_data_dictionary}
\alias{tcdsb_fetch_alchemer_data_dictionary}
\title{TCDSB Fetch Alchemer Survey Data
A function that retrieves data from TCDSB's Alchemer account.
Requires user to have set up an API with Alchemer, including
a token and secret key}
\usage{
tcdsb_fetch_alchemer_data_dictionary(survey_number)
}
\arguments{
\item{survey_number}{Alchemer survey number, retrieved from survey URL}
}
\value{
A data frame of question numbers and question labels
}
\description{
TCDSB Fetch Alchemer Survey Data
A function that retrieves data from TCDSB's Alchemer account.
Requires user to have set up an API with Alchemer, including
a token and secret key
}
\examples{
# token <- keyring::key_get("alchemer", "token")
# secret_key <- keyring::key_get("alchemer", "secret")
# tcdsb_fetch_alchemer_data_dictionary("XXXXXXXX")
}
