% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_ggplot_theme.R
\name{tcdsb_ggplot_theme}
\alias{tcdsb_ggplot_theme}
\title{TCDSB ggplot Theme}
\usage{
tcdsb_ggplot_theme(text_scale = 1)
}
\arguments{
\item{text_scale}{Numeric value to scale font sizes in plot. Default = 1.}
}
\value{
A ggplot theme
}
\description{
TCDSB ggplot Theme
}
\examples{
# df |>
#   ggplot(aes (x = x, y = y)) +
#   geom_bar() +
#   tcdsb_ggplot_theme()
}
