% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_gt_table.R
\name{tcdsb_gt_table}
\alias{tcdsb_gt_table}
\title{TCDSB gt tables}
\usage{
tcdsb_gt_table(
  df,
  table_font = "Century Gothic",
  fontsize = 1,
  title = "Title",
  subtitle = NA
)
}
\arguments{
\item{df}{A cleaned table for report}

\item{table_font}{Font}

\item{fontsize}{Font size scale - base is 1}

\item{title}{Title of table}

\item{subtitle}{Subtitle of table}
}
\value{
a gt object
}
\description{
TCDSB gt tables
}
\examples{
# df |>
#  tcdsb_gt_table(title = "Table for Report", font_size = 0.8)
}
