% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_presentation_template_quarto.R
\name{tcdsb_presentation_template_quarto}
\alias{tcdsb_presentation_template_quarto}
\title{Presentation Template - RevealJS}
\usage{
tcdsb_presentation_template_quarto(presentation_name)
}
\arguments{
\item{presentation_name}{File name}
}
\value{
Template to produce RevealJS presentation
}
\description{
Presentation Template - RevealJS
}
\examples{
# tcdsb_report_template_quarto(new_presentation)
}
