% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_report_template_pdf.R
\name{tcdsb_report_template_pdf}
\alias{tcdsb_report_template_pdf}
\title{Create a PDF Report from the TCDSB Template}
\usage{
tcdsb_report_template_pdf(name = "_example_report_pdf")
}
\arguments{
\item{name}{File name without extension. Defaults to "_example_report_pdf".}
}
\description{
Copy the example PDF report template into the current project,
optionally renaming it. Existing files are overwritten.
}
