% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_presentation_template_quarto.R,
%   R/tcdsb_report_template_quarto.R
\name{tcdsb_report_template_quarto}
\alias{tcdsb_report_template_quarto}
\title{Presentation Template - ReavelJS}
\usage{
tcdsb_report_template_quarto(report_name)

tcdsb_report_template_quarto(report_name)
}
\arguments{
\item{report_name}{Name of the report to be created.}

\item{presentation_name}{File name}
}
\value{
Template to produce RevealJS presentation

Templates and scripts to create the report
}
\description{
Presentation Template - ReavelJS

Report Template - Quarto PDF
A function that downloads an .qmd template and associate .typ and .lua files for a TCDSB themed report template.
}
\examples{
# tcdsb_report_template_quarto(new_presentation)
#tcdsb_report_template_quarto("test_report")
}
