% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_report_template_word.R
\name{tcdsb_report_template_word}
\alias{tcdsb_report_template_word}
\title{Report Template - Word
A function that downloads an .Rmd template, .R script and Word templates for a TCDSB themed report template.}
\usage{
tcdsb_report_template_word(report_name)
}
\arguments{
\item{report_name}{Name of the report to be created.}
}
\value{
Templates and scripts to create the report
}
\description{
Report Template - Word
A function that downloads an .Rmd template, .R script and Word templates for a TCDSB themed report template.
}
