% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_retrieve_alchemer.R
\name{tcdsb_retrieve_alchemer}
\alias{tcdsb_retrieve_alchemer}
\title{TCDSB - Retrieve Alchemer Data}
\usage{
tcdsb_retrieve_alchemer(x)
}
\arguments{
\item{x}{Alchemer survey number}
}
\value{
Data frame with survey data, plus a data dictionary
}
\description{
TCDSB - Retrieve Alchemer Data
}
\examples{
# tcdsb_retrieve_alchemer("XXXXXXX")

}
