% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_colours_fonts.R
\name{tcdsb_colours_fonts}
\alias{tcdsb_colours_fonts}
\title{TCDSB Colors and Fonts}
\usage{
tcdsb_colours_fonts(...)
}
\arguments{
\item{...}{Leave blank.}
}
\value{
R objects with names and HEX colors for TCDSB visual identity
}
\description{
TCDSB Colors and Fonts
}
\examples{
# tcdsb_colours_fonts()
}
