% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_presentation_template_quarto.R
\name{tcdsb_presentation_template_quarto}
\alias{tcdsb_presentation_template_quarto}
\title{Create a RevealJS Presentation from the TCDSB Template}
\usage{
tcdsb_presentation_template_quarto(name = "_example_presentation")
}
\arguments{
\item{name}{Optional file name (without extension). Defaults to
"presentation".}
}
\description{
Copy the example RevealJS presentation into the current project,
optionally renaming it. Existing files are overwritten.
}
