% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_project_setup.R
\name{tcdsb_project_setup}
\alias{tcdsb_project_setup}
\title{TCDSB Project Setup
A function that sets up a project folder. Creates a README file subfolders.}
\usage{
tcdsb_project_setup()
}
\value{
README file and subfolders
}
\description{
TCDSB Project Setup
A function that sets up a project folder. Creates a README file subfolders.
}
\examples{
# tcdsb_project_setup()
}
