% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_report_template_pdf.R
\name{tcdsb_report_template_pdf}
\alias{tcdsb_report_template_pdf}
\title{Report Template - PDF
A function that downloads an .Rmd template, images and .tex files for a TCDSB themed report template.}
\usage{
tcdsb_report_template_pdf(report_name)
}
\arguments{
\item{report_name}{Name of the report to be created.}
}
\value{
Templates and images for a TCDSB themed report
}
\description{
Report Template - PDF
A function that downloads an .Rmd template, images and .tex files for a TCDSB themed report template.
}
\examples{
# tcdsb_report_template_pdf(test_report)
}
