% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcdsb_report_template_quarto.R
\name{tcdsb_report_template_quarto}
\alias{tcdsb_report_template_quarto}
\title{Report Template - Quarto PDF
A function that downloads an .qmd template and associate .typ and .lua files for a TCDSB themed report template.}
\usage{
tcdsb_report_template_quarto(report_name)
}
\arguments{
\item{report_name}{Name of the report to be created.}
}
\value{
Templates and scripts to create the report
}
\description{
Report Template - Quarto PDF
A function that downloads an .qmd template and associate .typ and .lua files for a TCDSB themed report template.
}
\examples{
#tcdsb_report_template_quarto("test_report")
}
